restructuring course//added surveyapp

+++ src/main/java/com/example/training/springboottesting/app/repo/AnswerRepository.java
@@ -4,6 +4,6 @@ import com.example.training.springboottesting.app.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer>, JpaSpecificationExecutor<Answer> {
+public interface AnswerRepository extends JpaRepository<Answer, Long>, JpaSpecificationExecutor<Answer> {
 
 }
