+++ src/main/java/com/example/training/springboottesting/app/repo/QuestionRepository.java
@@ -4,6 +4,6 @@ import com.example.training.springboottesting.app.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface QuestionRepository extends JpaRepository<Question, Integer>, JpaSpecificationExecutor<Question> {
+public interface QuestionRepository extends JpaRepository<Question, Long>, JpaSpecificationExecutor<Question> {
 
 }
